Move models fixtures in to `mavis.test`
This moves the fixtures related to the page model objects in to their
own file.
