+++ mavis/test/fixtures/__init__.py
@@ -1,3 +1,16 @@
+from .models import (
+    children_page,
+    consent_page,
+    dashboard_page,
+    import_records_page,
+    log_in_page,
+    programmes_page,
+    school_moves_page,
+    sessions_page,
+    start_page,
+    unmatched_page,
+    vaccines_page,
+)
 from .playwright import (
     base_url,
     basic_auth,
@@ -10,6 +23,17 @@ __all__ = [
     "base_url",
     "basic_auth",
     "browser_context_args",
+    "children_page",
+    "consent_page",
+    "dashboard_page",
+    "import_records_page",
+    "log_in_page",
     "playwright_operations",
+    "programmes_page",
+    "school_moves_page",
     "screenshots_path",
+    "sessions_page",
+    "start_page",
+    "unmatched_page",
+    "vaccines_page",
 ]
